Should have done older commits, Lot of stuff done
Login leads to 4 page carousel. The first ones done, but the second one will need to be redone so that it is signs of dollar bills and then sumbols for plus and minus there.
Third page needs to talk about how much the cashier told you that you will get back and it will verify that number with subtracting page 2 data with page 3 data